backend/requirements.txt
pylint
Werkzeug
Flask
Flask-SQLAlchemy
Flask-Cors
python-jose
python-dotenv
authlib
requests
Flask-Migrate 